refactored to use new settings. One settings for entire project
