+++ data/src/main/java/com/helpfulapps/data/api/weather/converter/WeatherAnalyzer.kt
@@ -2,11 +2,11 @@ package com.helpfulapps.data.api.weather.converter
 
 import com.helpfulapps.data.db.converter.DayWeatherConverter
 import com.helpfulapps.data.db.weather.model.DayWeather
-import com.helpfulapps.data.helper.Units
+import com.helpfulapps.domain.helpers.Settings
 import io.reactivex.Single
 
 //TODO maybe it should be wrapped in factory pattern ?
-fun Single<List<DayWeather>>.analyzeWeather(units: Units) =
+fun Single<List<DayWeather>>.analyzeWeather(units: Settings.Units) =
     this.map { dayWeatherList ->
         dayWeatherList.forEach { DayWeatherConverter.analyzeWeather(it, units) }
         dayWeatherList
