+++ device/src/main/java/com/helpfulapps/device/alarms/AppAlarmManager.kt
@@ -6,8 +6,8 @@ import android.content.Context
 import android.os.Build
 import com.helpfulapps.base.extensions.rx.completableOf
 import com.helpfulapps.device.alarms.helpers.IntentCreator
-import com.helpfulapps.device.alarms.helpers.SettingsDevice
 import com.helpfulapps.device.alarms.helpers.matchesVersionsFrom
+import com.helpfulapps.domain.helpers.Settings
 import com.helpfulapps.domain.helpers.TimeSetter
 import com.helpfulapps.domain.repository.AlarmClockManager
 import io.reactivex.Completable
@@ -17,7 +17,7 @@ import com.helpfulapps.domain.models.alarm.Alarm as DomainAlarm
 class AlarmClockManagerImpl(
     private val context: Context,
     private val manager: AlarmManager,
-    private val settings: SettingsDevice
+    private val settings: Settings
 ) :
     AlarmClockManager {
